Fix epochs should be int

+++ run2_exp_diff_batch_sizes.py
@@ -20,7 +20,7 @@ def main(debug):
 
     x_train, y_train, x_test, y_test, num_classes = prepare_data(debug)
     batches = [32, 128, 512, 20000, len(x_train)]
-    WEIGHT_UPDATES = len(x_train) / 32 * 100
+    WEIGHT_UPDATES = int(len(x_train) / 32 * 100)
 
     os.makedirs(models_dir, exist_ok=True)
     os.makedirs(plots_dir, exist_ok=True)
